Migrate the steps components to app/components to align with Phlex conversions
