--- app/views/components/steps/container.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Steps
-  class Container < ApplicationComponent
-    def view_template(&)
-      div(class: "space-y-4", &)
-    end
-  end
-end
